+++ src/presenters/pop-overs/team-user-info-and-remove-pop.jsx
@@ -1,16 +1,13 @@
-/* globals Set */
-
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import {getAvatarThumbnailUrl, getDisplayName} from '../../models/user';
-import {getAvatarUrl} from  '../../models/project';
+import {getAvatarThumbnailUrl} from '../../models/user';
 
-import PopoverNested, {NestedPopoverConsumer, NestedPopoverTitle} from './popover-nested.jsx';
+import PopoverNested, {NestedPopoverConsumer} from './popover-nested.jsx';
 import {UserLink} from '../includes/link.jsx';
-import Loader from '../includes/loader.jsx';
-import Notifications from '../notifications.jsx';
 import Thanks from '../includes/thanks.jsx';
+
+import TeamUserRemovePop from './team-user-remove-pop.jsx';
 
 const MEMBER_ACCESS_LEVEL = 20;
 const ADMIN_ACCESS_LEVEL = 30;
@@ -110,155 +107,14 @@ const TeamUserInfo = ({currentUser, ...props}) => {
 
 // Team User Remove 💣
 
-class TeamUserRemove extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      gettingUser: true,
-      userTeamProjects: [],
-      selectedProjects: new Set(),
-    };
-    this.selectAllProjects = this.selectAllProjects.bind(this);
-    this.unselectAllProjects = this.unselectAllProjects.bind(this);
-    this.handleCheckboxChange = this.handleCheckboxChange.bind(this);
-    this.removeUser = this.removeUser.bind(this);
-  }
-  
-  removeUser() {
-    this.props.togglePopover();
-    this.props.createNotification(`${getDisplayName(this.props.user)} removed from Team`);
-    this.props.removeUserFromTeam(this.props.user.id, Array.from(this.state.selectedProjects));
-  }
-  
-  selectAllProjects() {
-    this.setState(({userTeamProjects}) => ({
-      selectedProjects: new Set(userTeamProjects.map(p => p.id))
-    }));
-  }
-  
-  unselectAllProjects() {
-    this.setState({
-      selectedProjects: new Set()
-    });
-  }
-  
-  handleCheckboxChange(evt) {
-    const {checked, value} = evt.target;
-    this.setState(({selectedProjects}) => {
-      selectedProjects = new Set(selectedProjects);
-      if (checked) {
-        selectedProjects.add(value);
-      } else {
-        selectedProjects.delete(value);
-      }
-      return {selectedProjects};
-    });
-  }
-  
-  async getUserWithProjects() {
-    const {data} = await this.props.api.get(`users/${this.props.user.id}`);
-    this.setState({
-      userTeamProjects: data.projects.filter(userProj => {
-        return this.props.team.projects.some(teamProj => teamProj.id === userProj.id);
-      }),
-      gettingUser: false
-    });
-  }
-  
-  componentDidMount() {
-    this.getUserWithProjects();    
-  }
-  
-  render() {
-    const allProjectsSelected = this.state.userTeamProjects.every(p => this.state.selectedProjects.has(p.id));
-    const userAvatarStyle = {backgroundColor: this.props.user.color};
-    
-    let projects = null;
-    if (this.state.gettingUser) {
-      projects = <Loader/>;
-    } else if (this.state.userTeamProjects.length > 0) {
-      projects = (
-        <React.Fragment>
-          <p className="action-description">
-            Also remove them from these projects
-          </p>
-          <div className="projects-list">
-            { this.state.userTeamProjects.map(project => (
-              <label key={project.id} htmlFor={`remove-user-project-${project.id}`}>
-                <input className="checkbox-project" type="checkbox" id={`remove-user-project-${project.id}`}
-                  checked={this.state.selectedProjects.has(project.id)} value={project.id}
-                  onChange={this.handleCheckboxChange}
-                />
-                <img className="avatar" src={getAvatarUrl(project.id)} alt={`Project avatar for ${project.domain}`}/>
-                {project.domain}
-              </label>
-            ))}
-          </div>
-          {this.state.userTeamProjects.length > 1 && (
-            <button className="button-small"
-              onClick={allProjectsSelected ? this.unselectAllProjects : this.selectAllProjects}
-            >
-              {allProjectsSelected ? 'Unselect All' : 'Select All'}
-            </button>
-          )}
-        </React.Fragment>
-      );
-    }
-    
-    return (
-      <dialog className="pop-over team-user-info-pop team-user-remove-pop">
-        <NestedPopoverTitle>
-          Remove {getDisplayName(this.props.user)}
-        </NestedPopoverTitle>
-        
-        <section className="pop-over-actions" id="user-team-projects">
-          {projects || (
-            <p className="action-description">
-              {getDisplayName(this.props.user)} is not a member of any projects
-            </p>
-          )}
-        </section>
-        
-        <section className="pop-over-actions danger-zone">
-          <button className="button-small has-emoji" onClick={this.removeUser}>
-            Remove <img className="emoji avatar" src={getAvatarThumbnailUrl(this.props.user)} alt={this.props.user.login} style={userAvatarStyle}/>
-          </button>
-        </section>
-      </dialog>
-    );
-  }
-}
-
-TeamUserRemove.propTypes = {
-  api: PropTypes.func.isRequired,
-  user: PropTypes.shape({
-    name: PropTypes.string,
-    login: PropTypes.string,
-    thanksCount: PropTypes.number.isRequired,
-    isOnTeam: PropTypes.bool,
-    color: PropTypes.string,
-  }).isRequired,
-  team: PropTypes.shape({
-    projects: PropTypes.array.isRequired
-  }),
-  removeUserFromTeam: PropTypes.func.isRequired,
-  createNotification: PropTypes.func.isRequired,
-};
-
 
 // Team User Info or Remove
 // uses removeTeamUserVisible state to toggle between showing user info and remove views
 
 const TeamUserInfoAndRemovePop = (props) => (
-  <PopoverNested menu={() => (
-    <Notifications>
-      {notifyFuncs => (
-        <TeamUserRemove {...notifyFuncs} {...props}/>
-      )}
-    </Notifications>
-  )}>
-    {() => (
-      <TeamUserInfo {...props}/>
+  <PopoverNested menu={() => <TeamUserRemovePop {...props}/>}>
+    {showRemove => (
+      <TeamUserInfo {...props} showRemove={showRemove}/>
     )}
   </PopoverNested>
 );
@@ -268,7 +124,6 @@ TeamUserInfoAndRemovePop.propTypes = {
     name: PropTypes.string,
     login: PropTypes.string,
     thanksCount: PropTypes.number.isRequired,
-    isOnTeam: PropTypes.bool,
     color: PropTypes.string,
   }).isRequired,
   currentUserIsOnTeam: PropTypes.bool.isRequired,
@@ -284,9 +139,6 @@ TeamUserInfoAndRemovePop.propTypes = {
 };
 
 TeamUserInfoAndRemovePop.defaultProps = {
-  user: {
-    isOnTeam: false
-  },
   currentUserIsOnTeam: false,
 };
